Add --python-executable option to plot command
The python scripts cannot be run using Python 3 but in some systems the
default "python" command points to Python 3 and there is another command
(e.g. python2.7) for other versions of Python. This commit lets the user
define the correct executable to use for Python 2.

It would be good if the Python scripts could be version agnostic but
this might need more work.